requirements.txt
discord.py
google-genai
flask
pillow
requests